+++ source/renderer/app/App.js
@@ -14,6 +14,7 @@ import AboutDialog from './containers/static/AboutDialog';
 import DaedalusDiagnosticsDialog from './containers/status/DaedalusDiagnosticsDialog';
 import BlockConsolidationStatusDialog from './containers/status/BlockConsolidationStatusDialog';
 import GenericNotificationContainer from './containers/notifications/GenericNotificationContainer';
+import AdaRedemptionUpgradeOverlay from './components/wallet/ada-redemption/AdaRedemptionUpgradeOverlay';
 import { DIALOGS } from '../../common/ipc/constants';
 import type { StoresMap } from './stores/index';
 import type { ActionsMap } from './actions/index';
@@ -30,11 +31,15 @@ export default class App extends Component<{
   }
   render() {
     const { stores, actions, history } = this.props;
-    const { app } = stores;
-    const { isActiveDialog } = app;
-    const locale = stores.profile.currentLocale;
+    const { app, profile } = stores;
+    const { isActiveDialog, openExternalLink } = app;
     const mobxDevTools = global.environment.mobxDevTools ? <DevTools /> : null;
-    const { currentTheme } = stores.profile;
+    const {
+      closeUpgradeOverlay,
+      currentLocale: locale,
+      currentTheme,
+      shouldShowUpgradeOverlay,
+    } = profile;
     const themeVars = require(`./themes/daedalus/${currentTheme}.js`).default;
     const { ABOUT, BLOCK_CONSOLIDATION, DAEDALUS_DIAGNOSTICS } = DIALOGS;
     return (
@@ -48,6 +53,11 @@ export default class App extends Component<{
               <Fragment>
                 <Router history={history} routes={Routes} />
                 {mobxDevTools}
+                <AdaRedemptionUpgradeOverlay
+                  onCloseOverlay={closeUpgradeOverlay}
+                  onOpenExternalLink={openExternalLink}
+                  showOverlay={shouldShowUpgradeOverlay}
+                />
                 {isActiveDialog(ABOUT) && <AboutDialog />}
                 {isActiveDialog(BLOCK_CONSOLIDATION) && (
                   <BlockConsolidationStatusDialog />